# specs/001-kanban-board/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: Kanban Board API
  description: REST API for managing Kanban board tasks, labels, and assignees
  version: 1.0.0
  contact:
    name: Kanban Board Team

servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://kanban.example.com/api
    description: Production server

paths:
  /tasks:
    get:
      summary: Get all tasks with associated labels and assignees
      description: Returns complete board state including all tasks, labels, and assignees
      tags:
        - Tasks
      parameters:
        - name: search
          in: query
          description: Search text to filter tasks by title or description
          required: false
          schema:
            type: string
        - name: labelIds
          in: query
          description: Comma-separated label IDs to filter by
          required: false
          schema:
            type: string
        - name: assigneeIds
          in: query
          description: Comma-separated assignee IDs to filter by
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  labels:
                    type: array
                    items:
                      $ref: "#/components/schemas/Label"
                  assignees:
                    type: array
                    items:
                      $ref: "#/components/schemas/Assignee"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new task
      description: Creates a new task in the TODO column
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /tasks/{id}:
    patch:
      summary: Update an existing task
      description: Updates task properties including status, labels, and assignees
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Delete a task
      description: Permanently deletes a task and its associated data
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /tasks/reorder:
    post:
      summary: Reorder or move a task
      description: Changes task position within or between columns
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskReorderRequest"
      responses:
        "200":
          description: Task reordered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /tasks/{id}/history:
    get:
      summary: Get task status history
      description: Returns time tracking data for a task
      tags:
        - Tasks
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskHistory"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /labels:
    get:
      summary: Get all labels
      description: Returns all available labels
      tags:
        - Labels
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  labels:
                    type: array
                    items:
                      $ref: "#/components/schemas/Label"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: Create a new label
      description: Creates a new label with specified name and color
      tags:
        - Labels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabelCreateRequest"
      responses:
        "201":
          description: Label created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  label:
                    $ref: "#/components/schemas/Label"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Label with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /assignees:
    get:
      summary: Get all assignees
      description: Returns all available team members
      tags:
        - Assignees
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  assignees:
                    type: array
                    items:
                      $ref: "#/components/schemas/Assignee"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - status
        - position
        - version
        - createdAt
        - updatedAt
        - statusChangedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Fix login bug"
        description:
          type: string
          nullable: true
          maxLength: 2000
          example: "Users cannot log in with email addresses"
        status:
          $ref: "#/components/schemas/Status"
        position:
          type: integer
          minimum: 0
          example: 0
        version:
          type: integer
          minimum: 1
          example: 1
          description: Version number for optimistic locking
        createdAt:
          type: string
          format: date-time
          example: "2025-12-04T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-12-04T11:45:00Z"
        statusChangedAt:
          type: string
          format: date-time
          example: "2025-12-04T11:00:00Z"
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/Assignee"

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Implement dark mode"
        description:
          type: string
          maxLength: 2000
          example: "Add dark mode theme toggle"
        labelIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["label-123", "label-456"]
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["user-789"]

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        status:
          $ref: "#/components/schemas/Status"
        labelIds:
          type: array
          items:
            type: string
            format: uuid
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
        version:
          type: integer
          minimum: 1
          description: Current version for optimistic locking

    TaskReorderRequest:
      type: object
      required:
        - taskId
        - targetColumn
        - targetPosition
      properties:
        taskId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        targetColumn:
          $ref: "#/components/schemas/Status"
        targetPosition:
          type: integer
          minimum: 0
          example: 2

    TaskHistory:
      type: object
      required:
        - id
        - taskId
        - status
        - enteredAt
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/Status"
        enteredAt:
          type: string
          format: date-time
          example: "2025-12-04T10:00:00Z"
        exitedAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-12-04T12:00:00Z"
        duration:
          type: integer
          nullable: true
          description: Duration in seconds
          example: 7200

    Label:
      type: object
      required:
        - id
        - name
        - color
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "label-123"
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Bug"
        color:
          type: string
          pattern: "^#[0-9A-F]{6}$"
          example: "#EF4444"
          description: Hex color code from predefined palette
        createdAt:
          type: string
          format: date-time
          example: "2025-12-04T09:00:00Z"

    LabelCreateRequest:
      type: object
      required:
        - name
        - color
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          example: "Feature"
        color:
          type: string
          pattern: "^#[0-9A-F]{6}$"
          example: "#3B82F6"
          description: Must be from predefined palette

    Assignee:
      type: object
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: "user-789"
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Alice Johnson"
        email:
          type: string
          format: email
          nullable: true
          example: "alice@example.com"
        createdAt:
          type: string
          format: date-time
          example: "2025-12-01T08:00:00Z"

    Status:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - TESTING
        - DONE
      example: "TODO"

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "BadRequest"
        message:
          type: string
          example: "Task title is required"
        details:
          type: object
          additionalProperties: true

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "BadRequest"
            message: "Task title must be between 1 and 200 characters"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "NotFound"
            message: "Task not found"

    Conflict:
      description: Resource conflict (e.g., concurrent modification)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Conflict"
            message: "Task was modified by another user. Please refresh and try again."

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"

tags:
  - name: Tasks
    description: Task management operations
  - name: Labels
    description: Label management operations
  - name: Assignees
    description: Assignee management operations